clean up runtime
